Added CalculateHelperTests and DiscountHelperTests

+++ src/UnitTestTutorial/Helpers/CalculateHelper.cs
@@ -5,12 +5,12 @@ namespace UnitTestTutorial.Helpers
 {
     public class CalculateHelper : ICalculateHelper
     {
-        public double Add(int x, int y)
+        public double Add(double x, double y)
         {
             return x+y;
         }
 
-        public double Divide(int x, int y)
+        public double Divide(double x, double y)
         {
             if(y==0)
             {
@@ -19,12 +19,12 @@ namespace UnitTestTutorial.Helpers
             return x/y;
         }
 
-        public double Multiply(int x, int y)
+        public double Multiply(double x, double y)
         {
             return x*y;
         }
 
-        public double Subtract(int x, int y)
+        public double Subtract(double x, double y)
         {
             return x-y;
         }
